% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rindex.function.R
\name{rindex.function}
\alias{rindex.function}
\title{Interpolated Refractive Function}
\usage{
rindex.function(pageid)
}
\arguments{
\item{pageid}{The material pageid from rindex.search}
}
\value{
A list containing a function for n, k and the orginal dataset
}
\description{
Interpolated Refractive Function
}
\examples{
rindex.search("Ag")

Ag <- rindex.function(0)

wavelength <- seq(0.25,0.8,,500)

# Interpolated Functions
plot(wavelength,Ag$n(wavelength), type='l', ylim=c(0,5.5), xlab=expression(wavelength~(mu*m)), ylab="refractive index n/k", lwd=3)
lines(wavelength,Ag$k(wavelength), col=2, lwd=3)
# Actual Data Points
points(Ag$data$wavelength, Ag$data$n)
points(Ag$data$wavelength, Ag$data$k, col=2)
title("Silver (Ag)")
legend(0.25,5,c("n","k"), col=1:2, lwd=3)
}

