% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rindex.search.R
\name{rindex.search}
\alias{rindex.search}
\title{Search the Refractive Index Materials Database}
\usage{
rindex.search(material)
}
\arguments{
\item{material}{Search string for material search}
}
\value{
Dataframe of matched materials.
}
\description{
Search the Refractive Index Materials Database
}
\examples{

}

