% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rindex.function.R
\name{rindex.function}
\alias{rindex.function}
\title{Interpolated Refractive Function}
\usage{
rindex.function(pageid)
}
\arguments{
\item{pageid}{The material pageid from rindex.search}
}
\value{
A function that will return the complex refractive index(s), given the wavelength(s) in meters
}
\description{
Interpolated Refractive Function
}
\examples{
rindex.search("Ag")

Ag_data <- rindex.get(0)
Ag <- rindex.function(0)

wavelengths <- seq(210,12300,, 1000) * 1e-9

plot(Ag_data$wavelength, Ag_data$n, log="x",  xlab = "wavelength (m) - log scale", ylab = "n")
lines(wavelengths, Re(Ag(wavelengths)), col = 2, lwd = 2)
}
